expression: add explicit unknown value for expression handling
This adds an explicit unknown value that gets set when unmarshaling
config expressions that have references set on them.

This should help differentiate between actual missing values, and
nil/null JSON values.
