+++ config.go
@@ -1,7 +1,6 @@
 package tfjson
 
 import (
-	"encoding/json"
 	"errors"
 )
 
@@ -167,103 +166,4 @@ type ModuleCall struct {
 
 	// The version constraint for modules that come from the registry.
 	VersionConstraint string `json:"version_constraint,omitempty"`
-}
-
-// Expression describes the format for an individual key in a
-// Terraform configuration.
-//
-// This struct wraps ExpressionData to support custom JSON parsing.
-type Expression struct {
-	*ExpressionData
-}
-
-// ExpressionData describes the format for an individual key in a
-// Terraform configuration.
-type ExpressionData struct {
-	// If the *entire* expression is a constant-defined value, this
-	// will contain the Go representation of the expression's data.
-	ConstantValue interface{} `json:"constant_value,omitempty"`
-
-	// If any part of the expression contained values that were not
-	// able to be resolved at parse-time, this will contain a list of
-	// the referenced identifiers that caused the value to be unknown.
-	References []string `json:"references,omitempty"`
-
-	// A list of complex objects that were nested in this expression.
-	// If this value is a nested block in configuration, sometimes
-	// referred to as a "sub-resource", this field will contain those
-	// values, and ConstantValue and References will be blank.
-	NestedBlocks []map[string]*Expression `json:"-"`
-}
-
-// UnmarshalJSON implements json.Unmarshaler for Expression.
-func (e *Expression) UnmarshalJSON(b []byte) error {
-	result := new(ExpressionData)
-
-	// Check to see if this is an array first. If it is, this is more
-	// than likely a list of nested blocks.
-	var rawNested []map[string]json.RawMessage
-	if err := json.Unmarshal(b, &rawNested); err == nil {
-		result.NestedBlocks, err = unmarshalExpressionBlocks(rawNested)
-		if err != nil {
-			return err
-		}
-	} else {
-		// It's a non-nested expression block, parse normally
-		if err := json.Unmarshal(b, &result); err != nil {
-			return err
-		}
-	}
-
-	e.ExpressionData = result
-	return nil
-}
-
-func unmarshalExpressionBlocks(raw []map[string]json.RawMessage) ([]map[string]*Expression, error) {
-	var result []map[string]*Expression
-
-	for _, rawBlock := range raw {
-		block := make(map[string]*Expression)
-		for k, rawExpr := range rawBlock {
-			var expr *Expression
-			if err := json.Unmarshal(rawExpr, &expr); err != nil {
-				return nil, err
-			}
-
-			block[k] = expr
-		}
-
-		result = append(result, block)
-	}
-
-	return result, nil
-}
-
-// MarshalJSON implements json.Marshaler for Expression.
-func (e *Expression) MarshalJSON() ([]byte, error) {
-	// Check for nested blocks, and marshal those instead if they exist.
-	if len(e.ExpressionData.NestedBlocks) > 0 {
-		return marshalExpressionBlocks(e.ExpressionData.NestedBlocks)
-	}
-
-	return json.Marshal(e.ExpressionData)
-}
-
-func marshalExpressionBlocks(nested []map[string]*Expression) ([]byte, error) {
-	var rawNested []map[string]json.RawMessage
-	for _, block := range nested {
-		rawBlock := make(map[string]json.RawMessage)
-		for k, expr := range block {
-			raw, err := json.Marshal(expr)
-			if err != nil {
-				return nil, err
-			}
-
-			rawBlock[k] = raw
-		}
-
-		rawNested = append(rawNested, rawBlock)
-	}
-
-	return json.Marshal(rawNested)
 }
