+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/hashicorp/terraform-json
 
 go 1.12
+
+require github.com/davecgh/go-spew v1.1.1
