+++ test-fixtures/explicit_null/plan.json
@@ -1,0 +1,173 @@
+{
+  "format_version": "0.1",
+  "terraform_version": "0.12.0-dev",
+  "variables": {
+    "one": {
+      "value": "two"
+    }
+  },
+  "planned_values": {
+    "root_module": {
+      "resources": [
+        {
+          "address": "null_resource.bar",
+          "mode": "managed",
+          "type": "null_resource",
+          "name": "bar",
+          "provider_name": "null",
+          "schema_version": 0,
+          "values": {
+            "triggers": {
+              "foo": "two"
+            }
+          }
+        },
+        {
+          "address": "null_resource.baz",
+          "mode": "managed",
+          "type": "null_resource",
+          "name": "baz",
+          "provider_name": "null",
+          "schema_version": 0,
+          "values": {
+            "triggers": {
+              "foo": "two"
+            }
+          }
+        },
+        {
+          "address": "null_resource.foo",
+          "mode": "managed",
+          "type": "null_resource",
+          "name": "foo",
+          "provider_name": "null",
+          "schema_version": 0,
+          "values": {
+            "triggers": null
+          }
+        }
+      ]
+    }
+  },
+  "resource_changes": [
+    {
+      "address": "null_resource.bar",
+      "mode": "managed",
+      "type": "null_resource",
+      "name": "bar",
+      "change": {
+        "actions": [
+          "create"
+        ],
+        "before": null,
+        "after": {
+          "triggers": {
+            "foo": "two"
+          }
+        },
+        "after_unknown": {
+          "id": true,
+          "triggers": {
+            "foo": false
+          }
+        }
+      }
+    },
+    {
+      "address": "null_resource.baz",
+      "mode": "managed",
+      "type": "null_resource",
+      "name": "baz",
+      "change": {
+        "actions": [
+          "create"
+        ],
+        "before": null,
+        "after": {
+          "triggers": {
+            "foo": "two"
+          }
+        },
+        "after_unknown": {
+          "id": true,
+          "triggers": {
+            "foo": false
+          }
+        }
+      }
+    },
+    {
+      "address": "null_resource.foo",
+      "mode": "managed",
+      "type": "null_resource",
+      "name": "foo",
+      "change": {
+        "actions": [
+          "create"
+        ],
+        "before": null,
+        "after": {
+          "triggers": null
+        },
+        "after_unknown": {
+          "id": true,
+          "triggers": false
+        }
+      }
+    }
+  ],
+  "configuration": {
+    "root_module": {
+      "resources": [
+        {
+          "address": "null_resource.bar",
+          "mode": "managed",
+          "type": "null_resource",
+          "name": "bar",
+          "provider_config_key": "null",
+          "expressions": {
+            "triggers": {
+              "references": [
+                "var.one"
+              ]
+            }
+          },
+          "schema_version": 0
+        },
+        {
+          "address": "null_resource.baz",
+          "mode": "managed",
+          "type": "null_resource",
+          "name": "baz",
+          "provider_config_key": "null",
+          "expressions": {
+            "triggers": {
+              "references": [
+                "var.one"
+              ]
+            }
+          },
+          "schema_version": 0
+        },
+        {
+          "address": "null_resource.foo",
+          "mode": "managed",
+          "type": "null_resource",
+          "name": "foo",
+          "provider_config_key": "null",
+          "expressions": {
+            "triggers": {
+              "constant_value": {
+                "foo": null
+              }
+            }
+          },
+          "schema_version": 0
+        }
+      ],
+      "variables": {
+        "one": {}
+      }
+    }
+  }
+}
