+++ test-fixtures/explicit_null/terraform.tfvars
@@ -1,0 +1,1 @@
+one = "two"
